fix TS issues within jest service worker + fix typos
