+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "video-editor",
+  "name": "visual-editor",
   "version": "0.1.0",
   "private": true,
   "type": "module",
@@ -58,7 +58,6 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "@types/react-modal": "^3.16.3",
-    "@types/service-worker-mock": "^2.0.4",
     "@types/uuid": "^10.0.0",
     "address": "^2.0.3",
     "eslint": "^9",
